# src/main/resources/static/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is spec for our BOOK API",
    "version" : "v1",
    "title" : "Swagger spec for our BOOK API"
  },
  "host" : "localhost:8080",
  "basePath" : "/Book_API",
  "tags" : [ {
    "name" : "Book API"
  } ],
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/books" : {
      "post" : {
        "tags" : [ "Book API" ],
        "summary" : "Creates a book",
        "description" : "",
        "operationId" : "create",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Book"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Failure"
          }
        }
      }
    },
    "/books/{id}" : {
      "get" : {
        "tags" : [ "Book API" ],
        "summary" : "Retrieves a book based on ID.",
        "description" : "",
        "operationId" : "read",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Book"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Failure"
          }
        }
      }
    }
  },
  "definitions" : {
    "Book" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "title" : {
          "type" : "string"
        },
        "author" : {
          "type" : "string"
        }
      }
    }
  }
}